Update rootScatterPlot.c
Was able to fix the issue of the g TGraph not displaying by moving the declaration before the for-loops and inserting the filling and drawing of the g TGraph (including the updating of the can1 TCanvas) into the nested for-loop (before incrementing i). Also, changed the declaration of y_phi, x_eta, and i to formal data types.